feat: remove test code from nr contracts, add new ts tests for exploit
