+++ noir-projects/noir-contracts/contracts/token_contract/src/main.nr
@@ -214,23 +214,6 @@ contract Token {
         Token::at(context.this_address()).assert_minter_and_mint(context.msg_sender(), amount).enqueue(&mut context);
     }
 
-    // Showing an exploit
-    #[aztec(private)]
-    fn privately_mint_private_note_to(amount: Field, to: AztecAddress) {
-        storage.balances.add(to, U128::from_integer(amount)).emit(encode_and_encrypt_note(&mut context, to, to));
-
-        Token::at(context.this_address()).assert_minter_and_mint(context.msg_sender(), amount).enqueue(&mut context);
-    }
-
-    #[aztec(private)]
-    fn privately_mint_private_note_to_direct(amount: Field, to: AztecAddress) {
-        // Acts as an entrypoint, so we need to end the setup phase
-        context.end_setup();
-        storage.balances.add(to, U128::from_integer(amount)).emit(encode_and_encrypt_note(&mut context, to, to));
-
-        Token::at(context.this_address()).assert_minter_and_mint(context.msg_sender(), amount).enqueue(&mut context);
-    }
-
     #[aztec(public)]
     #[aztec(internal)]
     fn assert_minter_and_mint(minter: AztecAddress, amount: Field) {
