+++ yarn-project/end-to-end/src/e2e_crowdfunding_and_claim.test.ts
@@ -9,11 +9,13 @@ import {
   Fr,
   Note,
   type PXE,
+  PackedValues,
+  TxExecutionRequest,
   type TxHash,
   computeSecretHash,
   deriveKeys,
 } from '@aztec/aztec.js';
-import { computePartialAddress } from '@aztec/circuits.js';
+import { GasSettings, TxContext, computePartialAddress } from '@aztec/circuits.js';
 import { InclusionProofsContract } from '@aztec/noir-contracts.js';
 import { ClaimContract } from '@aztec/noir-contracts.js/Claim';
 import { CrowdfundingContract } from '@aztec/noir-contracts.js/Crowdfunding';
@@ -359,6 +361,47 @@ describe('e2e_crowdfunding_and_claim', () => {
     await expect(
       claimContract.withWallet(donorWallets[0]).methods.claim(note, donorWallets[0].getAddress()).send().wait(),
     ).rejects.toThrow();
+  });
+
+  it('cannot withdraw as non operator', async () => {
+    const donationAmount = 500n;
+
+    // 1) We add authwit so that the Crowdfunding contract can transfer donor's DNT
+    const action = donationToken
+      .withWallet(donorWallets[1])
+      .methods.transfer_from(donorWallets[1].getAddress(), crowdfundingContract.address, donationAmount, 0);
+    const witness = await donorWallets[1].createAuthWit({ caller: crowdfundingContract.address, action });
+    await donorWallets[1].addAuthWitness(witness);
+
+    // 2) We donate to the crowdfunding contract
+    await crowdfundingContract.withWallet(donorWallets[1]).methods.donate(donationAmount).send().wait({
+      debug: true,
+    });
+
+    // Calling the function normally will fail as msg_sender != operator
+    await expect(
+      crowdfundingContract.withWallet(donorWallets[1]).methods.withdraw(donationAmount).send().wait(),
+    ).rejects.toThrow('Assertion failed: Not an operator');
+
+    // Instead, we construct a call and impersonate operator by skipping the usual account contract entrypoint...
+    const call = crowdfundingContract.withWallet(donorWallets[1]).methods.withdraw(donationAmount).request();
+    // ...using the withdraw fn as our entrypoint
+    const entrypointPackedValues = PackedValues.fromValues(call.args);
+    const request = new TxExecutionRequest(
+      call.to,
+      call.selector,
+      entrypointPackedValues.hash,
+      new TxContext(donorWallets[1].getChainId(), donorWallets[1].getVersion(), GasSettings.default()),
+      [entrypointPackedValues],
+      [],
+    );
+    // NB: Removing the msg_sender assertion from private_init will still result in a throw, as we are using
+    // a non-entrypoint function (withdraw never calls context.end_setup()), meaning the min revertible counter will remain 0.
+    // This does not protect fully against impersonation as the contract could just call context.end_setup() and the below would pass.
+    // => the private_init msg_sender assertion is required (#7190, #7404)
+    await expect(donorWallets[1].simulateTx(request, true, operatorWallet.getAddress())).rejects.toThrow(
+      'Assertion failed: Users cannot set msg_sender in first call',
+    );
   });
 
   it('cannot donate after a deadline', async () => {
